# docs/documentai-api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Document AI API",
    "description": "API for document processing",
    "version": "v1"
  },
  "paths": {
    "/v1/documents": {
      "post": {
        "tags": [
          "Documents:Upload"
        ],
        "summary": "Create Document",
        "description": "Upload a document for processing (fire-and-forget).",
        "operationId": "create_document_v1_documents_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-Trace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_document_v1_documents_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadAsyncResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/wait": {
      "post": {
        "tags": [
          "Documents:Upload"
        ],
        "summary": "Create Document Wait",
        "description": "**Deprecated.** Upload a document and poll until processing completes or timeout.\n\nThis endpoint is deprecated and will be removed in a future release.\nAPI Gateway HTTP API enforces a hard 30-second integration timeout; BDA extraction\ncommonly takes 30-50 seconds, so this endpoint cannot reliably return a result before\nthe gateway terminates the connection.\n\nUse `POST /v1/documents` instead and poll `GET /v1/documents/{job_id}` for the result.",
        "operationId": "create_document_wait_v1_documents_wait_post",
        "deprecated": true,
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "include_extracted_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Extracted Data"
            }
          },
          {
            "name": "include_bounding_box",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Bounding Box"
            }
          },
          {
            "name": "timeout",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 25,
              "title": "Timeout"
            }
          },
          {
            "name": "X-Trace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_create_document_wait_v1_documents_wait_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/{job_id}": {
      "get": {
        "tags": [
          "Documents:Query"
        ],
        "summary": "Get Document Results",
        "description": "Get processing results by job ID.",
        "operationId": "get_document_results_v1_documents__job_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          },
          {
            "name": "include_extracted_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Extracted Data"
            }
          },
          {
            "name": "include_bounding_box",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Include Bounding Box"
            }
          },
          {
            "name": "X-Trace-ID",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Trace-Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents:Delete"
        ],
        "summary": "Delete Document",
        "description": "Delete a document by job ID.\n\nsoft_delete=True (default): retain the S3 files, mark the record DELETED\n(recoverable). soft_delete=False: also purge every S3 copy of the document -\noriginal upload, preprocessing copies, and BDA output (hard delete).",
        "operationId": "delete_document_v1_documents__job_id__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Job Id"
            }
          },
          {
            "name": "soft_delete",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Soft Delete"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/documents/search": {
      "post": {
        "tags": [
          "Documents:Query"
        ],
        "summary": "Search Documents",
        "description": "Search for multiple documents by job IDs.",
        "operationId": "search_documents_v1_documents_search_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DocumentSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentSearchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_create_document_v1_documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[a-z0-9 _-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "Type of document being uploaded"
          },
          "external_document_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "pattern": "^[\\w.\\-/]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Document Id",
            "description": "External document identifier"
          },
          "external_system_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "pattern": "^[\\w.\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "External System Id",
            "description": "External system identifier"
          },
          "ai_consent_flag": {
            "type": "boolean",
            "title": "Ai Consent Flag",
            "description": "Explicit AI processing consent. True = consent granted, processing proceeds normally. False = consent denied, document is stored but not processed.",
            "default": true
          },
          "upload_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadSource"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source device type of the upload (desktop or mobile)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_document_v1_documents_post"
      },
      "Body_create_document_wait_v1_documents_wait_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "category": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64,
                "minLength": 1,
                "pattern": "^[a-z0-9 _-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category",
            "description": "Type of document being uploaded"
          },
          "external_document_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 256,
                "pattern": "^[\\w.\\-/]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Document Id",
            "description": "External document identifier"
          },
          "external_system_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128,
                "pattern": "^[\\w.\\-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "External System Id",
            "description": "External system identifier"
          },
          "ai_consent_flag": {
            "type": "boolean",
            "title": "Ai Consent Flag",
            "description": "Explicit AI processing consent. True = consent granted, processing proceeds normally. False = consent denied, document is stored but not processed.",
            "default": true
          },
          "upload_source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UploadSource"
              },
              {
                "type": "null"
              }
            ],
            "description": "Source device type of the upload (desktop or mobile)"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_create_document_wait_v1_documents_wait_post"
      },
...
      "DocumentSearchRequest": {
        "properties": {
          "jobIds": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Jobids"
          },
          "includeExtractedData": {
            "type": "boolean",
            "title": "Includeextracteddata",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "jobIds"
        ],
        "title": "DocumentSearchRequest"
      },
      "DocumentSearchResponse": {
        "properties": {
          "results": {
            "items": {
              "$ref": "#/components/schemas/JobStatusResponse"
            },
            "type": "array",
            "title": "Results"
          }
        },
        "type": "object",
        "required": [
          "results"
        ],
        "title": "DocumentSearchResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UploadAsyncResponse": {
        "properties": {
          "jobId": {
            "type": "string",
            "title": "Jobid"
          },
          "jobStatus": {
            "type": "string",
            "title": "Jobstatus"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "jobId",
          "jobStatus",
          "message"
        ],
        "title": "UploadAsyncResponse"
      },
      "UploadSource": {
        "type": "string",
        "enum": [
          "desktop",
          "mobile"
        ],
        "title": "UploadSource"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "API-Key"
      },
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}